Bluestein FFT convolution converted to Cooley-Tukey FFT convolution

+++ src/ffte.h
@@ -1,10 +1,10 @@
 #ifndef FFTE_H_
 #define FFTE_H_
 
-void ffteshift(float x_real[], float x_imag[], size_t N, unsigned char only_real_input);
-void iffteshift(float x_real[], float x_imag[], size_t N, unsigned char only_real_input);
+void ffteshift(double x_real[], double x_imag[], size_t N, unsigned char only_real_input);
+void iffteshift(double x_real[], double x_imag[], size_t N, unsigned char only_real_input);
 
-void ffte(float x_real[], float x_imag[], size_t N, unsigned char only_real_input);
-void iffte(float x_real[], float x_imag[], size_t N, unsigned char only_real_input);
+void ffte(double x_real[], double x_imag[], size_t N, unsigned char only_real_input);
+void iffte(double x_real[], double x_imag[], size_t N, unsigned char only_real_input);
 
 #endif
